dtype float 64

+++ datasets/ReefDataset.py
@@ -90,11 +90,13 @@ class ReefDataset(Dataset):
             transformed = self.transforms(image=np.array(img),
                                           bboxes=target['boxes'],
                                           class_labels=target['labels'])
-            img = torch.as_tensor(transformed['image'], dtype=torch.float32)
+            img = torch.as_tensor(transformed['image'], dtype=torch.float64)
             target["boxes"] = torch.as_tensor(transformed['bboxes'])
             target["labels"] = torch.as_tensor(transformed['class_labels'])
+        
         print(img)
         print(target)
+
         return img, target
 
     def __len__(self):
